+++ app/src/main/java/lab/justonebyte/moneysubuu/model/Transaction.kt
@@ -8,7 +8,7 @@ enum class TransactionType(val value:Int){
     Expense(2)
 }
 
-data class Transaction(val id:Int?=null,val amount:Double,val type: TransactionType,val category: TransactionCategory,val created_at:Double){
+data class Transaction(val id:Int?=null,val amount:Int,val type: TransactionType,val category: TransactionCategory,val created_at:Double){
     object  Mapper {
         fun mapToDomain(transactionEntity: TransactionWithCategory):Transaction{
             return Transaction(
